Implement ScanNet DataLoader

+++ test.py
@@ -39,7 +39,7 @@ class Test:
         for step, batch_data in enumerate(self.data_loader):
             # Get the inputs and labels
             inputs = batch_data[0].to(self.device)
-            labels = batch_data[1].to(self.device)
+            labels = batch_data[1].long().to(self.device)
 
             with torch.no_grad():
                 # Forward propagation
